Add fractional item pickup
A bit of a mess and hacky but it works

+++ BobDerKleinkriminelle/Assets/Scripts/ItemData.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "ItemData", menuName = "ItemData", order = 1)]
 public class ItemData : ScriptableObject {
     public string itemName;
-    public float value;
-    public float weight;
+    public int value;
+    public int weight;
     //public Vector3 itemPosition;
 }
